applets/notifications: don't take length of string twice
To determine whether there are any actions, the notification applet
takes the length of some property that may be undefined, then replaces
the result with an empty string if falsy, and takes the length again.
This will cause an error if the original value is undefined, and produce
an undefined value for later comparison if it's not.

Instead, don't take the length immediately and only after the guard.