applets/notifications: restore padding above the icon, when visible
We removed row spacing from the base GridLayout, resulting in content
on the first row touching the header. This is fine for text which
includes internal top padding, but it's not great for the icon which
often will not, depending on what exactly the app asks for. Restore the
top padding for just the icon.


(cherry picked from commit 44411ce43e7ec939ac68715138ae86452bc99ff2)

4cf6f850 applets/notifications: restore padding above the icon, when visible

Co-authored-by: Nate Graham <nate@kde.org>