screenpool: Remove incorrect assert
The documentation for outputOrder states:

```
At the time of change signalling all Screens are
guaranteed to exist in QGuiApplication.

After screen removal this can temporarily contain
dead entries.
```

aka:
Anything in qApp::screens  will be in outputOrder
Not everything in outputOrder will be in qApp::screens

This assert doesn't match the documentation and invents new
assumptions.

The actualy code is written with this in mind and there is a matching
guard in ScreenPool.