applets/digital-clock: don't use Qt.font()
The font object for the custom font is currently created using the
Qt.font() method. This will cause an error if there is no argument and
requires some actual values for matching a font; an empty object will
not work either.

Instead, we can just set it to null; qml will create a font object when
the property is set.