applets/digitalclock: only prompt if calendar plugins are different
The set of enabled calendar plugins are not stored with a cfg_ property,
but managed directly by the configuration page through the saveConfig
function and the configurationChanged signal. This means that changing
then restoring the plugin state will still consider the configuration
modified and prompt the user to save or discard it.

Instead, check if the plugin state is actually different from the saved
configuration, and update the unsavedChanges property accordingly.