plasmacalendarintegration: only prompt user if config different
The holiday plugin configuration uses the configurationChanged signal
to inform the dialog that changes need to be saved. This will stay even
if the user manually changes the configuration back to the previous
settings. The applet configuration dialog can now also use the
unsavedChanges property to be informed more dynamically if the
configuration needs to be modified.

This change implements this method for the holiday calendar plugin.