kcms/wallpaper: set roleNames for WallpaperConfigModel
WallpaperConfigModel, which contains the available wallpaper plugins, is
a PlasmaQuick::ConfigModel. This class has the capability to load a full
kcm with name stored under the PluginNameRole through its KCMRole, which
was used previously to load kcms as  applet configuration pages (but
this seems to no longer be actively used by applets we ship). When this
model is assigned to the plugin selection combobox, the combobox creates
a copy of the currently active entry from the model, including the
KCMRole. In other words, it pointlessly tries to load a kcm with the
same name as the wallpaper plugin, which fails because it's not a kcm
but a wallpaper plugin. This does nothing except cause an error, and if
a kcm with that name happened to exist, it would only waste resources to
instantiate it here because we don't do anything with it.

We can solve this by overriding the roleNames and removing the KCMRole;
this way the combobox will not try to access it as it doesn't know it
exists.