kcms/wallpaper: set wallpaper properties initially
755123b3 changed wallpaper plugin loading to set the configuration
properties after loading the plugin rather than initially to solve the
large number of warnings caused by configuration keys that are not used
by the plugin. This works without issues for our own plugins, but can
cause issues with some third-party plugins that do unusual things (such
as serializing complex configuration as json that's saved in the plasma
config file). This should not be difficult to fix in those plugins, but
we should not unnecessarily break compatibility within one major Plasma
release.
Instead, we load a temporary version of the configuration and use it to
figure out the initial property values for the real configuration. This
solves the warnings while maintaining backward-compatibility.