applets/digital-clock: set auto toolbar for time zone config
configTimeZones unlike other applet configuration pages is not a
KCMUtils component, but a nested PageRow to allow some fancy handling
for the map overlay. With the Breadcrumb header this directly works, but
switching to a Toolbar (Auto) header breaks this as it does not force
the header over the internal PageRow, leading to an empty header.

Instead, we set the auto header here as well, which fixes the issue.