applets/digital-clock: move Add button to page header
The time zone currently has its own header, which only contains the Add
button. It feels a little lost there, sandwiched between the list and
some helper text, and is not really recognizable as a ListView header.

Instead place the button where we would typically place it in a kcm: in
the page header.