applets/systray: use shared visibility combobox model
The combobox model for visibility is created by a function in each
delegate. Before disabling SNIs was possible, this was necessary as the
model was different based on whether the entry was a SNI or a plasmoid.

Now that the model no longer depends on the type, we no longer need this
and can just use a single model for all of them (either the full or a
reduced one depending on the state of the "Always show all" setting.


Co-authored-by: Ismael Asensio <isma.af@gmail.com>