applets/digital-clock: move usage tip to footer
The time zone configuration currently has a usage tip at the very top.
This is a rather unusual position, we usually have these below the
element that it is a tip for.

This change moves the usage tip to the footer, aligns it with the
checkbox, and sets it to small font to match the other helper text.