applets/digital-clock: adjust usage tip wording
The time zone configuration page has a good usage tip, but the wording
is not optimal. It hedges its own usefulness with "If you're frequently
traveling", which is not necessary as it's also useful if only traveling
to different time zones occasionally. It does not explain why the user
might want to do this, and is a bit unclear in parts.

This change rephrases this label.

Co-authored-by: Tem PQD <variable_valuables761@simplelogin.com>