applets/kicker: Highlight newly installed apps
This tracks installed apps in kickerstaterc and keeps them
highlighted in the menu until launched for a maximum of
3 days. The "All Apps" category doesn't highlight since the
idea of this feature is to teach the user *where* to find
the app.

A timer checks for expired highlights every 24 ours.
The model updates could surely be optimized but they're rare,
so the added complexity isn't worth it imho.

CCBUG: 481426