applets/digitalclock: don't show time zone sidebar with only one tz
We advertise that users should add their home timezone as an extra time
zone so it'll be there when traveling, and that this will not cause it
to show up unnecessarily. That's not true though; if the user has local
and their home timezones configured, it will show the time zone sidebar
with a single entry in the fullRepresentation popup, as the visibility
check uses the full list of timezones, not the deduplicated one.

Instead, use the deduplicated list.