applets/notifications: enable DND while fullscreen app is focused
Resolves https://invent.kde.org/plasma/plasma-workspace/-/issues/119.

There is currently a setting to allow normal-priority notifications to
appear in full-screen windows, which doesn't actually affect the DND mode.
This change replaces that setting with a more intuitive behavior: while a
full-screen window is focused, DND mode is enabled, and when the focus is
no longer on the full-screen window, DND mode is disabled. Benefits:

- Notifications appearing or not is less magic, and the user can easily
  understand why they are appearing or not.
- DND mode is actually activated when inhibiting notifications, which
  allows the rest of the system to know that the user is busy with
  something, and apps can read the `org.freedesktop.Notifications.Inhibited`
  property to check the status of DND mode.