digital-clock: emit signal when setting selected time zones
When setting the list of selected time zones for the time zone model,
as the config page does when adding a new time zone, the corresponding
Changed() signal is not sent, so the config page does not realize that a
change was made and the apply button stays disabled (until something
else triggers a check, when the difference in settings is detected).

Instead, emit the Changed() signal so the Apply button is enabled
immediately.