applets/systemtray: warn people about disabling klipper
When you do this, you lose all clipboard history, including keeping
copied stuff when the app it came from closes. This is non-obvious, and
I've seen a number of confused people complaining about it over the
years.

Add a warning about this so people will at least have the opportunity to
learn that what they're doing is not recommended.