ksmserver: Drop subsession management
KSMServer has support for subsessions in order to support stopping
activities.

This feature is a flawed design that does not work and could never work.
Apps can be on mulitple activities, or be a single instance application
with an instance already running when resumed.

In combination with XSMP being blocked on sandboxed apps this is even
more broken. At this point we should simplify the code.