applets/mediacontroller: allow arrow keys on slider after opening
When the mediacontroller popup is opened, keyboard focus is on the
representation item itself. Until 6cb910de, left/right arrow keys did
nothing there until the user pressed down arrow to pass focus to a real
item. That commit only allowed passing focus to the tab bar though,
leaving left/right arrow still dead in the default state if only one
player is open (and also had less-than-optimal focus handling, not
actually putting activeFocus on the tab itself, so pressing down would
still focus it first).

This change actually passes focus to the element, and extends this to
the slider so left/right arrow keys do something whether the tab bar is
active or not. Along the way, we allow arrow up from the tab bar to
focus the header in the system tray version, for consistency with other
widgets (like Audio Volume).