klipper/declarative: reorder DelegateToolButtons
The action tool buttons on the widget/popup delegates are currently in
this order: Actions, QR Code, Edit, Delete. This is not optimal: Actions
is highly useful to a subset of users, but needs manual setup and more
knowledge about how this feature works (or knowing that we ship URL
actions by default, which is completely hidden from the UI). QR Code is
useful only in rather niche circumstances. Edit is relatively useful,
easy to understand, and does not require setup. Delete clearly needs to
be last for consistency with other delegates.

Instead, sort the (first three) actions by how useful they are to a
general user base: Edit, Action, QR Code, Delete.