dataengines/weather: Use KHolidays/SunEvents to check for night time
Use the new functionality provided by `KHolidays::SunEvents` to check
whether a specific observation at a location corresponds to day or
night time.

This ports away from using the legacy `time` dataengine and massively
simplifies the code, as the check is now synchronous.

It applies to bbcukmet, noaa, and envcan providers. DWD already got
this information from the provider response