applets/DigitalClock: add bottom separator to map
Our view-based applet config pages usually have separator lines at the
top and bottom of the view. The map mode of the Time Zone page does not,
and replaces the default button pane that provides a separator (and also
does not use a KCMUtils base component, which also provide separator
functionality.

Instead, add a separator background to the fake footer similar to the
one on the real footer.