applets/systray: improve tab navigation in hiddenItemsView
Tab navigation in the hidden items popup sort of works, but suffers from
two major flaws: the focused item is not actually highlighted, and on
most entries there are additional pointless tab stops. This is very
confusing. The major cause is the several nested focusscopes with
activeFocusOnTab; any attempt to make it work more reasonably is
frustrating and extremely fragile.

Instead, we just drop activeFocusOnTab (only) while things are in the
hidden items popup, and keep a single tab stop for the full list.
This is simple, robust, and works well with screen readers. We also set
a custom description on the grid to properly announce the
two-dimensional structure over screen readers.


Co-authored-by: Nate Graham <nate@kde.org>