wallpapers/image: Fix wallpaper thumbnails not adjusting to output size
Amends d6436e077c05abd100d905aaa73222a48ca1695f.

This fixes wallpaper thumbnails not reloading after the screen size changes
(it's likely a caching issue somewhere in the model, the relevant code
is not easy to follow), it takes a load of complexity off the image model,
and it paves a way for future simplifications in the model.

With the proposed approach, there is a dedicated image provider for
wallpaper previews. The only thing that the model does is construct urls
that are fed to Image. If the screen size changes, the qml code will
change the Image.sourceSize property and the Image will do the rest for
us.


(cherry picked from commit cddbf3f2e23886c3871eb864412c57483c033ee4)

Co-authored-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>