kcms/fonts: warn if trying to use a "Display" font
These fonts are not actually intended for use on computer screens, but
the word "display" seems to cause people to make the opposite
assumption. As a result, there's a constant trickle of bug reports and
complaints about weird text when using one of these fonts.

Add a warning in the KCM so people learn the truth.