fix: notify about missed notifications only once
There is an issue where the summary about notifications missed while in
"Do Not Disturb" mode is shown unconditionally so long as those
notifications are unread, when we come out of Do Not Disturb mode.

This means that if the user toggles Do Not Disturb mode on and off
multiple times, they will see the summary multiple times, even if it was
shown only seconds ago.

This has become apparent since we are now properly using Do Not Disturb
for fullscreen windows, and so this bug causes the summary to be shown
every time we come out of fullscreen mode (even for a moment) if there
are unread notifications that were delivered while in Do Not Disturb
mode.