🍒 Notifications: Set notifications to read when full representation is opened
When full representation is set expanded, we should set all the items as
read,
since that is what the user is expecting to do when clicking on the
notification icon and it resets to normal.

If notifs are not set as read, then notifs that appeared during
Do-not-disturb will keep notifying the user, even if user has explicitly
looked at the items in the widget.

BUG: 504350
FIXED-IN: 6.4

(cherry picked from commit c5eda9ab9de64a5fd9225e4c9718422f72d10672)

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>