applets/kicker: Remember uninstalled apps for a short period When an installed .desktop file is updated, it might be removed and immediately re-added, which causes it to be inappropriately marked New! as the model is updated for both changes.
To resolve this, when an application is uninstalled, we remember it for three days. If it is still new, we also remember the date it became new. If the application is re-installed within 3 days, it will be restored with its previous new-ness. After that, as before, it would be new again.

This fixes both cases when an application is updated (as in the bug), as well as the case where a user might renege on a decision to uninstall an application.

BUG: 504523