kcms/componentchooser: fix dialog scrollability
I broke this in 46e9e9e469153df9b492855fed69ea520ca0df16 while porting
the old OverlaySheet version to Dialog; by overriding the contentItem
of the dialog, I blew away the invisible internal ScrollView that the
content gets magically re-parented to. Which is probably a good example
of why this kind of magic is a bad idea overall.

BUG: 504418
FIXED-IN: 6.3.6