libtaskmanager: show Comment for launcher icons when appropriate
Right now, launcher icons get a caption equal to the app's GenericName,
if one is set in the metadata. If not, it gets no caption.

This differs from what Kicker and Kickoff Manager do: a more complex
behavior to show the Comment if the app has no GenericName or if its
GenericName is equal to its Name. Let's replicate the same logic here.

BUG: 504431
FIXED-IN: 6.4.0