startkde: Set _JAVA_AWT_WM_NONREPARENTING environment variable
AWT has a hard coded list of non-reparenting window managers. If a Java
application misdetects that the window manager reparents windows, it can
reject ConfigureNotify events and paint nothing on the screen.

Recently, kwin_wayland had been changed so it doesn't reparent windows
anymore. This broke some Java applications because of hardcoded WM
behaviors in AWT. There is no "nice" way to fix it, but an environment
variable can be set to instruct AWT that it's a non-reparenting window
manager.

BUG: 504692