applets/digital-clock: use ListView for plugins
The calendar plugin names are not completely self-explanatory; if you
don't already know that "Alternate Calendar" shows the date in an
alternate calendar format or that "Calendar Events" is for showing
events from an Akonadi calendar, it might not be apparent from the
(otherwise sensible) names. The plugins already have description fields
in their metadata  (though the current strings are not quite optimal, as
they weren't exposed in the UI) and even icons, but we don't use them.

Instead, use a ListView and (Check)IconTitleSubtitle to display nice-
looking list entries for the plugins that can have descriptive labels
associated with them.

CCBUG: 504755