plasmashell: don't duplicate old applet configs in panel cloning
KConfigGroup::copyTo is recursive in sub groups, but in this case we don't want
to copy the sub groups representing the applets, rather to create new applets
because we want new ids.
copying the whole config might result in duplicated applets between the old and the new
panel.

so copy only the immediate panel config keys

BUG:504856