plasmacalendarintegration: make plugin description more informative
The calendar plugin descriptions are currently not exposed to users, but
we're preparing to do this to make the functionality of the plugins
clearer. The current description has some unnecessary bits (that it is a
plugin) and could instead be a bit more explicit.

This change updates the plugin description.