kcms/wallpaper: don't load PreviewImage
PreviewImage is a temporary internal key that shouldn't be stored, but
was (and will still be present in existing configs, as the Containment
dialog wrote it to disk). kcm_wallpaper does not write it, but if it is
present in the file, it may mark the kcm as needing save to remove it.

Instead, don't consider the key at all.

BUG: 504934