wallpapers/image: clear Preview key before saving
Image sets a preview key, which overrides the current main wallpaper,
intended for use in the Desktop containment config dialog to immediately
preview the currently selected wallpaper. This is written to the config
so in effect the preview is mostly used instead of the real image
setting, which is pointless.

Instead, we reset it to the default (string "null") right when saving;
this way the key does not end up in the file.

CCBUG: 504934


(cherry picked from commit d4ca36ad788e0ca43f8f9fce521f1e6d91fd531d)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>