kcms/fonts: save people from excessively small font sizes
The Qt font dialog lets you set the font to be absurdly small, down to
0.01pt.

Sizes this small don't make sense; let's set a lower bound of 4pt with
an appropriate warning message about that.

Let's also warn people if they try to set a size below 7pt, which is
not in the realm of "unusable" but rather "non-text UI elements may
all look too large". Instead, steer people towards the KScreen KCM, the
way we do if they try to use huge fonts.

BUG: 505175
FIXED-IN: 6.5.0