applets/kicker: make clear invokable from qml
Currently, kicker only allows setting the query to the empty string,
which will clear the query as well, but operates on a 10ms timer. This
makes it hard to programmatically reset the search in response to e.g.
newly installed applications - emptying and resetting fails as by the
time the query is forwarded to the runners, it's the same as the old
query and the cleared state never reached them.

Instead, we allow calling the existing clear method from qml.