applets/kicker: don't emit layoutChanged while looping over AppEntry
AppEntry's reload method always emits the owning model's layoutChanged
signal when called. This method is only called in two spots: when the
sycoca data changes, and when the name display settings are changed. In
both cases, we are iterating over several, potentially many items. This
can cause severe performance problems in particular with favorites in
kicker, which use a repeater that has to destroy and re-add all the
items in response to this signal. Installing or removing software can
cause a mad frenzy where each favorite causes the whole sidebar to be
rebuilt. This happens in-process, and with many favorites and several
kicker instances can freeze plasmashell for minutes while it pointlessly
destroys and recreates delegates.

Instead, we don't emit layoutChanged in the loop, but move it up to the
caller and just emit it once after we're done reloading all the items.


(cherry picked from commit db0ebc2014508f1605337a6f81b54cf79ec7f469)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>