applets/kicker: show all apps in history
4399573d (in 2015) introduced a filter in the history model to not show
favorites in recently used applications. In principle, this makes sense
(and apparently was requested at the time, though I could not find the
relevant feature requests) - as favorites tend to be often used,
including them makes the recent application history mostly a less easily
accessed copy of the favorites list. But this is opaque to users and
even developers, who often perceive these omissions as a bug, and there
does not seem to be a concise and clear wording that would make the
intentionality of this omission apparent - labeling menus or sections
"Recent Applications excluding Favorites" is terribly awkward.

Instead, be less clever and just do the obvious thing that the labels
History and Recent Applications suggest: show the history of the most
recently launched applications.

BUG: 435356
BUG: 449834