dataengines/weather: Adapt to the new dynamic URLs on envcan
The envcan provider is switching from using fixed urls for the
weather reports to dynamic urls based on their timestamp.

To find the exact name of the file we need to first parse some html
directory listings, which can be done reusing the xml parser and
some small tricks.

With the intent of minimizing the amount of changes, this patch
adds the new API as a fallback, if the original static file is
not found.


(cherry picked from commit 5d09beb22abb6217a1c5a0bb540ab817087295f8)

Co-authored-by: Ismael Asensio <isma.af@gmail.com>