kcms/lookandfeel: Encapsulate lnf application in manager more
Currently, one part of the look and feel application logic lives in the
kcm, while another lives in the LookAndFeelManager. From the design point
of view, that doesn't look great.

This change cleans up the code a bit so the LookAndFeelManager is solely
responsible for applying look and feel packages. It makes the look and
feel manager more reusable.