applets/kicker: resolve non-existing files to their url
Kicker's PAStats-based favorites do some normalization to, among others,
resolve symlinked files to their resolved path to reduce duplicates.
However, if the file doesn't exist, it essentially just gives up. This
is a problem when trying to remove favorites that were deleted: while
the un-favoriting in PAStats succeeds, kicker can't remove it from its
own model because it can't identify the file by its url if it doesn't
exist. So removed favorites confusingly stick around until the next
plasmashell restart.

Instead, we take the base url in that case - as the entries already have
gone through the process, the url is their real url anyway.

BUG: 402820
FIXED-IN: 6.4.1