KlipperPopup: Always start from beginning of the list
Currently we set the listView currentIndex to 0 when the view is hidden.
This causes the index to be wrong when the view is opened again, if
anything was copied when the view was hidden.

Instead just always set the currentIndex to 0, to make sure the first
item is always selected.

BUG: 505493
FIXED-IN: 6.4.1


(cherry picked from commit 44ca2410b471f7eed581c6dbf713f0c85af92054)

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>