kcms/feedback: remove little hover icons
These icons have an unclear purpose since they have no text, and the
same icons are not repeated elsewhere in the KCM's UI to provide some
clue as to their meaning or purpose. They only appear on hover and
can't be accessed on touch at all; let's just remove them to prevent
user confusion.

The backend code providing icons in the model is preserved in case we
ever want to do something else with icons in the future.

BUG: 505761
FIXED-IN: 6.4.1