weather/envcan: Fix handling of non-continuous hourly folder list
The previous code assumed the provider would always have a continuous
list of hourly folders that contain the forecast files. This is the
case most of the times, but I found an instance where this list had
an odd '23' folder leftover from the day before.

This was probably a small bug on the provider's side, but it would
make the forecast fail for users in that region the whole day.

Let's handle this in a more robust way, by storing the whole sorted
list of folders and following it along.