shell: Create panel views after desktop views are ready
With the current code, the splash screen will be hidden as soon as the
wallpapers are loaded.

However, the splash screnn is actually notified about the desktop stage
about 1-1.5 second later after the wallpaper plugin resets the loading
property.

The reason for that is that the panel is loaded between
`wallpaper.loading = false` in the wallpaper and ShellCorona::checkAllDesktopsUiReady().

This change re-arranges the startup sequence so the panels are loaded
after the desktop views become ready. It reduces plasma startup time a bit.

In long term, we should look for making panel loading as async as
possible so the main thread doesn't get blocked for too long.