kcms/notifications: Do not show System Notifications page for apps
When launching the KCM, the arguments can specify an application
(or service) to focus in, which we pass on to the SourcesPage via
the `rootIndex` property.

For sandboxed apps, which use the notifications portal, the origin
(notifyrc file) is actually `plasma_workspace` with an event-id
of type `notification`.

Check for this to avoid showing the System Notifications when we
actually want to configure the notifications for the app

Also we need to set these properties declaratively on the call to
open the page on the first load. Otherwise, the root index will get
overriden by the kcm arguments, showing a broken page

BUG: 506236
FIXED-IN: 6.4.2