kcms/notifications: Fix broken "configure notification" page
When clicking the configure button on a notification, we launch the
KCM with the notification info so it directly navigates to the
specific app/service and event.

When the notification event belongs to plasma-workspace we try to
show the System Notifications index, but if there is also a desktop
entry, the page will override this index to the application's one,
showing a "broken" page.

Fix this by setting all the desired properties on the call that
opens the Sources Settings page, instead of calculating it after
the page is loaded.

CCBUG: 506236
(cherry picked from commit 62f4c616c0a848b6f052f209a10f22ddba46f79c)