kcms/notifications: Rework how arguments are passed to QML
The initial arguments passed to the KCM are used just one time to
make the QML navigate to the specified app/service and event, so
the user can directly edit the notification they're interested in.

Instead of using context kcm properties that need to be reset, pass
those directly in the signal to QML, so the codeflow is straighforward
clear and enforces the single-shot intent.