applet/digital-clock: back to using sizehelper to calc width
applet/digital-clock: back to using sizehelper to calc width

This partially reverts:
https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/3533
https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/3538

as I believe the problem they tried to fix was due to wrong timeFormat used
in sizehelper, so we should fix that instead of not using sizehelper for width.
The later results in the applet constantly changing width with a variable-width
font.