fix(startkde): make plasma-ksmserver.service Wants instead of Requires (BUG: 503120)
### Reason for the change

This changes `plasma-ksmserver.service` from a strong dependency (`Requires=`) to a weak dependency (`Wants=`) in `plasma-core.target`. This means that, if ksmserver fails (which will happen, for example, in a Wayland session with Xwayland disabled), it will not cause the whole Plasma session to fail. (Currently this failure seems to only be triggered when the user service manager is reloaded, for example by running `systemctl --user daemon-reload`; I think this has to do with the order the services are loaded in.)

Note that `plasma-workspace-x11.target` has `BindsTo=plasma-ksmserver.service`, a strictly stronger dependency requirement, so this change has no impact on X11 sessions.

### Bugs fixed

BUG: 503120