servicerunner: drop favorites scores
while well intentioned, this feature is really just making the results
unreliable. it's not obvious to the user that a favorite should impact
search, making for odd sorting for no apparent reason and even when it
is apparent the favorite may still be a worse search result.

a result that recently came up was that `kate` could be ordered before
`code` while being a clearly not as good a match

(the code also lacked testing :(()