servicerunner: fuzzy match
use a bitap implementation instead of doing awkward contains dances.
this should lead to somewhat more reliable results, which are now more
comprehensively asserted in the unit test

at the heart of this is a new fuzzyScore function that assigns a score
to a service vis a vis a query. this score is adjusted depending on
which field it is regarding (name > genericname > keywords).
this should hopefully ensure that a match against name outweighs most
other matches. all scores are eventually assembled into a final score
that gets used as match relevance