Don't break desktop ordering when new desktop is added
Without this patch, any new created desktop instance has its position
set to 0. Since this is evaluated later to eventually sort, we
systematically end up with more than one desktop at position 0.

Meanwhile, we had the intended position for the desktop directly
available in the function call, so store that position in the newly
added desktop.

This avoids the confusion in the pager which shows new desktops wrongly
placed when added.