Reparent menu actions to the temporarily created QMenu in global menu
Otherwise, the search action will try to manipulate the menu of the
search action, but it will only change the empty one, and not the one
that we are displaying. By reparenting we make sure that all changes
are applied to the shown menu.

BUG:505876


(cherry picked from commit 5f18319f3bf15c5d92321473850b3fbae6a2f5b5)

Co-authored-by: Niccolò Venerandi <niccolo@venerandi.com>