applets/systemtray: Show background apps as reported by org.freedesktop.BackgroundMonitor
Some apps rely on the shell showing this info - they just vanish
when clicking the window but keep on running and continue for example
to play music. Without us displaying the app somewhere this is very
confusing to the user and there exist no other way to bring them
back short of "launching" the app gain which is not very intuitive
either.
Note that only flatpaks are currently reported through this mechanism.
A proxy model also makes sure that we do not display a duplicate
entry for an app if it itself registers a SNI.
Co-authored-by: Jin Liu <m.liu.jin@gmail.com>