wallpapers/image: Follow Plasma::Theme colors instead of system color
The wallpaper is a part of plasma and its expected that it follows
the plasma color instead of apps colors in "twilight" like set ups.
If the panel is dark and the applets that are on top of the wallpaper
are dark the wallpaper itself should also be the dark version.