libklookandfeel: New helper library
This tiny library provides shared helpers for determining whether a global
theme is light or dark. It is used by the look and feel and landing page
configuration modules.

In the future, the library can include other global theme related features.
For example, perhaps the look and feel manager could live in it too.