runners: reduce unnecessary usage of minimum letter counts
Users generally expect search results to start appearing with the first
character typed; we shouldn't be setting minimum letter counts without
a very good reason. Excessive minimum letter counts are also bad
specifically for Chinese, which is highly information-dense.

Get rid of most of the minimum letter counts, and reduce from 3 to 2 in
cases where this was done to prevent unnecessary DBus or Baloo queries,
which can be more expensive.

CCBUG: 490972